server/requirements.txt
celery==4.0.2
Django==1.9.1
django-cors-headers==1.1.0
django-filter==0.11.0
djangorestframework==3.3.2
django-rest-swagger==0.3.5
psycopg2==2.6.2
Markdown==2.6.5
redis==2.10.5
requests==2.9.1
sendgrid==2.2.1
sendgrid-django==2.0.0
